refactor : PointStatusRandoeGenerator는 Util에 가까우므로 패키지를 옮기다
